requirements.txt
# Core Django
Django==6.0.4
asgiref==3.11.1
sqlparse==0.5.5

# Database
dj-database-url==2.3.0
psycopg2-binary==2.9.10

# Server
gunicorn==23.0.0

# Static files
whitenoise==6.9.0

# REST & GraphQL
djangorestframework==3.17.1
strawberry-graphql==0.314.3
strawberry-graphql-django==0.82.1
strawberry-django-auth==0.379.1
graphql-core==3.2.8

# Auth & Security
PyJWT==2.12.1

# CORS
django-cors-headers==4.9.0

# Config
python-decouple==3.8

# Media
pillow==12.2.0

# Cache & Queues
redis==7.4.0
django-redis==6.0.0
celery==5.6.3
kombu==5.6.2
amqp==5.3.1
billiard==4.2.4
vine==5.1.0

# Celery CLI deps
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
prompt_toolkit==3.0.52
wcwidth==0.6.0

# HTTP
requests==2.33.1
urllib3==2.6.3
certifi==2026.2.25
charset-normalizer==3.4.7
idna==3.11

# Utilities
python-dateutil==2.9.0.post0
six==1.17.0
tzdata==2026.1
tzlocal==5.3.1
typing_extensions==4.15.0
packaging==26.1
cross-web==0.6.0
librt==0.9.0
